+++ spec/dummy/app/jobs/elastic_beanstalk_periodic_task.rb
@@ -3,5 +3,5 @@
 class ElasticBeanstalkPeriodicTask < ApplicationJob
   queue_as :default
 
-  def perform(); end
+  def perform; end
 end